+++ api/urls.py
@@ -14,6 +14,9 @@ router.register(r"importtasks", views.ImportTaskViewSet, basename="importtask")
 urlpatterns = [
     path("", views.APIOverview.as_view(), name="overview"),
     path("localhost-redirect", views.LocalHostRedirectView.as_view(), name="localhost"),
+    path(
+        "organisations/", views.OrganisationListView.as_view(), name="organisation-list"
+    ),
     path("gmn/", include(("gmn.urls", "gmn"), namespace="gmn")),
     path("gmw/", include(("gmw.urls", "gmw"), namespace="gmw")),
 ]